Use Fay-Text instead of custom Text
Since Yesod-Fay already depends on Fay-Text now,

https://github.com/faylang/fay-text/issues/1

we can use Fay-Text instead of the cusom Text, so we use the same Text
type everywhere, and can use the functions defined in Fay-Text as well.

Still needs a bit of work, toString is missing in Fay-Text. I left it
here for now, but it looks a bit strange.

Already modified Fay-DOM to depend on Fay-Text instead of String

https://github.com/faylang/fay-dom/pull/4 .